fix #3 采用markdown编辑器，初步实现页面静态化（友情链接、导航栏）

+++ src/main/java/cn/edu/sdufe/cms/common/dao/agency/AgencyMapper.java
@@ -26,6 +26,6 @@ public interface AgencyMapper extends GenericDao<Agency, Long> {
      * @param id
      * @return
      */
-    int updateViews(Long id);
+    long updateViews(Long id);
 
 }
